Fix Checkstyle warnings
