+++ domains/java/oop/calculating-volume/Display.java
@@ -1,0 +1,5 @@
+class Display {
+  void display(double x) {
+    System.out.printf("%.3f\n", x);
+  }
+}
